+++ test/knxip_tests/tunnelling_request_test.py
@@ -1,7 +1,5 @@
 """Unit test for KNX/IP TunnellingRequest objects."""
-import asyncio
-import unittest
-
+import pytest
 from xknx import XKNX
 from xknx.dpt import DPTBinary
 from xknx.exceptions import CouldNotParseKNXIP
@@ -10,19 +8,8 @@ from xknx.telegram import GroupAddress, Telegram
 from xknx.telegram.apci import GroupValueWrite
 
 
-class Test_KNXIP_TunnelingReq(unittest.TestCase):
+class TestKNXIPTunnelingRequest:
     """Test class for KNX/IP TunnellingRequest objects."""
-
-    # pylint: disable=too-many-public-methods,invalid-name
-
-    def setUp(self):
-        """Set up test class."""
-        self.loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(self.loop)
-
-    def tearDown(self):
-        """Tear down test class."""
-        self.loop.close()
 
     def test_connect_request(self):
         """Test parsing and streaming connection tunneling request KNX/IP packet."""
@@ -53,17 +40,14 @@ class Test_KNXIP_TunnelingReq(unittest.TestCase):
         knxipframe = KNXIPFrame(xknx)
         knxipframe.from_knx(raw)
 
-        self.assertTrue(isinstance(knxipframe.body, TunnellingRequest))
-        self.assertEqual(knxipframe.body.communication_channel_id, 1)
-        self.assertEqual(knxipframe.body.sequence_counter, 23)
-        self.assertTrue(isinstance(knxipframe.body.cemi, CEMIFrame))
+        assert isinstance(knxipframe.body, TunnellingRequest)
+        assert knxipframe.body.communication_channel_id == 1
+        assert knxipframe.body.sequence_counter == 23
+        assert isinstance(knxipframe.body.cemi, CEMIFrame)
 
-        self.assertEqual(
-            knxipframe.body.cemi.telegram,
-            Telegram(
-                destination_address=GroupAddress("9/0/8"),
-                payload=GroupValueWrite(DPTBinary(1)),
-            ),
+        assert knxipframe.body.cemi.telegram == Telegram(
+            destination_address=GroupAddress("9/0/8"),
+            payload=GroupValueWrite(DPTBinary(1)),
         )
 
         cemi = CEMIFrame(xknx, code=CEMIMessageCode.L_DATA_REQ)
@@ -76,14 +60,14 @@ class Test_KNXIP_TunnelingReq(unittest.TestCase):
         )
         knxipframe2 = KNXIPFrame.init_from_body(tunnelling_request)
 
-        self.assertEqual(knxipframe2.to_knx(), list(raw))
+        assert knxipframe2.to_knx() == list(raw)
 
     def test_from_knx_wrong_header(self):
         """Test parsing and streaming wrong TunnellingRequest (wrong header length byte)."""
         raw = (0x06, 0x10, 0x04, 0x20, 0x00, 0x15, 0x03)
         xknx = XKNX()
         knxipframe = KNXIPFrame(xknx)
-        with self.assertRaises(CouldNotParseKNXIP):
+        with pytest.raises(CouldNotParseKNXIP):
             knxipframe.from_knx(raw)
 
     def test_from_knx_wrong_header2(self):
@@ -91,5 +75,5 @@ class Test_KNXIP_TunnelingReq(unittest.TestCase):
         raw = (0x06, 0x10, 0x04, 0x20, 0x00, 0x15, 0x04)
         xknx = XKNX()
         knxipframe = KNXIPFrame(xknx)
-        with self.assertRaises(CouldNotParseKNXIP):
+        with pytest.raises(CouldNotParseKNXIP):
             knxipframe.from_knx(raw)